almost done with profile.js

+++ client/src/components/dashboard/Dashboard.js
@@ -4,13 +4,13 @@ import { Link } from 'react-router-dom';
 import { getCurrentProfile, DeleteWholeAccount } from '../../Actions/profileActions'
 import D_mainFrame from './D_mainFrame'
 
-// import Spinner from '../layouts/Spinner'
+import Spinner from '../layouts/Spinner'
 
 const Dashboard = ({ getCurrentProfile, DeleteWholeAccount, auth, profile }) => {
 
     useEffect(() => {
         getCurrentProfile(auth.token)
-    }, [])
+    }, [getCurrentProfile])
 
     const { user } = auth
 
@@ -37,37 +37,41 @@ const Dashboard = ({ getCurrentProfile, DeleteWholeAccount, auth, profile }) => 
     } 
     
     
-    return (
-      <Fragment>
-        <div>
-          <h1 className="large text-primary">Dashboard</h1>
-          <p className="lead">
-            <i className="fas fa-user" /> Welcome {user && user.name}
-          </p>
-          {profile.profile !== null ? (
-            <Fragment> {renderButtons()} </Fragment>
-          ) : (
-            <Fragment>
-              {profile.error.msg} <br />{" "}
-              <Link to="/create-profile" className="btn btn-primary">
-                Create Profile{" "}
-              </Link>{" "}
-            </Fragment>
-          )}
-        </div>
-          {profile.profile === null ? '' :  
-          <D_mainFrame
-          experience={!profile.profile ? [] : profile.profile.experience}
-          education={!profile.profile ? [] : profile.profile.education}
-        />}
-        <div className="my-2">
-          <button className="btn btn-danger" onClick={() => DeleteWholeAccount(auth.token)}>
-            <i className="fas fa-user-minus" />
-            Delete My Account
-          </button>
-        </div>
-      </Fragment>
-    );
+    if (auth.loading || profile.loading) {
+      return <Spinner />
+    } else {
+      return (
+        <Fragment>
+          <div>
+            <h1 className="large text-primary">Dashboard</h1>
+            <p className="lead">
+              <i className="fas fa-user" /> Welcome {user && user.name}
+            </p>
+            {profile.profile !== null ? (
+              <Fragment> {renderButtons()} </Fragment>
+            ) : (
+              <Fragment>
+                {profile.error && profile.error.msg} <br />{" "}
+                <Link to="/create-profile" className="btn btn-primary">
+                  Create Profile{" "}
+                </Link>{" "}
+              </Fragment>
+            )}
+          </div>
+            {profile.profile === null ? '' :  
+            <D_mainFrame
+            experience={!profile.profile ? [] : profile.profile.experience}
+            education={!profile.profile ? [] : profile.profile.education}
+          />}
+          <div className="my-2">
+            <button className="btn btn-danger" onClick={() => DeleteWholeAccount(auth.token)}>
+              <i className="fas fa-user-minus" />
+              Delete My Account
+            </button>
+          </div>
+        </Fragment>
+      );
+    }
 }
 
 const mapStateToProps = state => ({
